+++ exabel_data_sdk/query/query.py
@@ -8,14 +8,23 @@ from exabel_data_sdk.query.table import Table
 
 @dataclass
 class Query:
+    """
+    Represents a SELECT query to the data API.
+    """
 
+    # The table to retrieve data from
     table: Table
+
+    # The columns to retrieve
     columns: Sequence[Column]
-    filters: Sequence[Predicate]
+
+    # The WHERE part of the query
+    predicates: Sequence[Predicate]
 
     def sql(self) -> str:
+        """Returns the query string"""
         cols = ", ".join([col.sql() for col in self.columns])
         query = f"SELECT {cols} FROM {self.table.name}"
-        if self.filters:
-            query += " WHERE " + " AND ".join([f.sql() for f in self.filters])
+        if self.predicates:
+            query += " WHERE " + " AND ".join([p.sql() for p in self.predicates])
         return query
