+++ exabel_data_sdk/query/table.py
@@ -3,7 +3,6 @@ from dataclasses import dataclass
 
 @dataclass
 class Table:
-
     """The name of the table"""
 
     name: str
